Reorganize internal tasks
