+++ plugin/src/main/kotlin/com/github/triplet/gradle/play/tasks/PublishProducts.kt
@@ -1,9 +1,9 @@
 package com.github.triplet.gradle.play.tasks
 
 import com.github.triplet.gradle.play.internal.PRODUCTS_PATH
-import com.github.triplet.gradle.play.internal.PlayPublishTaskBase
 import com.github.triplet.gradle.play.internal.isDirectChildOf
 import com.github.triplet.gradle.play.internal.playPath
+import com.github.triplet.gradle.play.tasks.internal.PlayPublishTaskBase
 import com.google.api.client.json.jackson2.JacksonFactory
 import com.google.api.services.androidpublisher.model.InAppProduct
 import org.gradle.api.tasks.InputDirectory
